+++ example/integration_test/plugin_integration_test.dart
@@ -1,11 +1,3 @@
-// This is a basic Flutter integration test.
-//
-// Since integration tests run in a full Flutter application, they can interact
-// with the host side of a plugin implementation, unlike Dart unit tests.
-//
-// For more information about Flutter integration tests, please see
-// https://flutter.dev/to/integration-testing
-
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
@@ -14,11 +6,12 @@ import 'package:ss_preventer/ss_preventer.dart';
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
-  testWidgets('getPlatformVersion test', (WidgetTester tester) async {
-    final SsPreventer plugin = SsPreventer();
-    final String? version = await plugin.getPlatformVersion();
-    // The version string depends on the host platform running the test, so
-    // just assert that some non-empty string is returned.
-    expect(version?.isNotEmpty, true);
+  testWidgets('preventOn/preventOff does not throw', (
+    WidgetTester tester,
+  ) async {
+    await SsPreventer.preventOn();
+    await SsPreventer.preventOff();
+
+    expect(true, isTrue);
   });
 }