+++ lib/src/screenshot_event.dart
@@ -1,0 +1,5 @@
+class ScreenshotEvent {
+  ScreenshotEvent({required this.detectedAt});
+
+  final DateTime detectedAt;
+}